fix qr wallet route
